requirements.txt
dj-database-url==0.4.1
Django==1.9.8
django-crispy-forms==1.6.0
django-filter==0.14.0
django-geoposition==0.3.0
django-guardian==1.4.5
djangorestframework==3.4.6
gunicorn==19.6.0
Markdown==2.6.6
psycopg2==2.6.2
pytz==2016.7
six==1.10.0
whitenoise==3.2
